Move new positions to a realtivePosition utility file
